Added lpthread dependency

+++ src/main.cpp
@@ -9,7 +9,7 @@ using namespace std;
 void list_containers()
 {
     vector<Container> containers = Docker::list_containers();
-    for(int i = 0; i < containers.size(); i++)
+    for(long unsigned int i = 0; i < containers.size(); i++)
     {
         printf("%s %s\n", containers[i].get_id().c_str(), containers[i].get_name().c_str());
     }
@@ -18,7 +18,7 @@ void list_containers()
 void list_images()
 {
     vector<Image> images = Docker::list_images();
-    for(int i = 0; i < images.size(); i++)
+    for(long unsigned int i = 0; i < images.size(); i++)
     {
         printf("%s %s\n", images[i].get_id().c_str(), images[i].get_repository().c_str());
     }
@@ -33,7 +33,7 @@ int main(int argc, const char * argv[])
 
     string content = "";
     vector<string> vec = Utils::read_file("src/views/index.html");
-    for(int i = 0; i < vec.size(); i++)
+    for(long unsigned int i = 0; i < vec.size(); i++)
     {
         content += vec[i];
     }
